Add files via upload
This is the Software Requirements Specification (SRS) file for the project of the Telemedicine Website.
prepared by CSE327.6 - Project Group 7.

SRS: A software requirements specification (SRS) is a document that describes what the software will
do and how it will be expected to perform. An SRS describes the functionality the product needs to fulfill all stakeholders (business, users)
needs.